+++ source/extensions/lightspeed.trex.stage_manager.plugin.tree.usd/lightspeed/trex/stage_manager/plugin/tree/usd/mesh_groups.py
@@ -85,7 +85,11 @@ class MeshGroupsModel(_VirtualGroupsModel):
                     )
                 )
 
-        return list(tree_items.values())
+        # Sort the items alphabetically (both parents and children)
+        sorted_tree_items = list(tree_items.values())
+        self.sort_items(sorted_tree_items)
+
+        return sorted_tree_items
 
 
 class MeshGroupsDelegate(_VirtualGroupsDelegate):